backend/requirements.txt
ollama
pdfplumber 
langchain 
langchain-core
langchain-ollama
langchain-community
langchain_text_splitters
unstructured
unstructured[all-docs]
fastembed
pikepdf
sentence-transformers
elevenlabs
streamlit
python-docx
langchain_classic
langchain-docling
langchain_milvus
pymilvus
easyocr

fastapi
starlette
python-multipart
uvicorn[standard]

motor
pymongo
pydantic
pydantic[email]

pyvi



pyjwt
itsdangerous
pwdlib[argon2]
httpx
python-dotenv
